refact: Expression 객체 domain 모듈로 이동
